src/crawler/requirements.txt
PyYAML==6.0.2
python-dotenv==1.0.1
pymongo==4.10.1
aiohttp==3.10.11
requests==2.32.3
beautifulsoup4>=4.12
lxml>=4.9
